Port opcode 0xDB and repair ActorData pointer widths
Port func_80091F84 from asm and convert ActorData's unk110-unk120 pointer run to u32, matching the FieldActor convention, so sizeof(ActorData) returns to the retail 0x138 and the raw PSX offsets used by the actor init and model render paths stay correct on 64-bit hosts.

The widened fields let func_80080A74's raw-offset stores poison unk118 (0xffff00000000), which the new 0xDB handler was first to dereference. Named users reconstruct host pointers via uintptr_t casts; the ActorData allocation now uses sizeof(ActorData), and the script VM's 0x138 snapshot copy is exactly sized again.

+++ src/field/scripts/virtual_machine.c
@@ -445,7 +445,7 @@ void func_800A2714(void) {
             if (pActor->unk124 != -1) {
                 g_FieldScriptVMCurActor = pActor;
                 pData = HeapAlloc(ArchiveDecodeAlignedSize(pActor->unk124, pActor) + 8, 0x0);
-                g_FieldScriptVMCurActor->unk120 = pData;
+                g_FieldScriptVMCurActor->unk120 = (u32)(uintptr_t)pData;
                 ArchiveReadFileToBuffer(g_FieldScriptVMCurActor->unk124, pData, 0, CdlModeSpeed); // Read from disc into buffer
                 ArchiveCdDataSync(0);
             }
